pyproject.toml
[tool.poetry]
name = "doitwithvriska"
version = "0.1.0"
description = ""
authors = ["leo60228 <leo@60228.dev>"]
license = "GPL-2.0"

[tool.poetry.dependencies]
python = "^3.8"
2captcha-python = "^1.0.3"
requests = "^2.25.1"
Faker = "^5.6.5"
python-dotenv = "^0.15.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
